Improvements / Changes to support proving out DO extension - cluster hooks optional - buff up save hook (and pass in cluster when calling apply fn) - move normalizeName into generic place - bring back async create machine config - hack for do extension (map example provider to do provider)
