+++ docusaurus/docs/extensions/home.md
@@ -7,3 +7,7 @@ Extensions can be created independently of Rancher and their code can live in se
 You can find some example extensions here: https://github.com/rancher/ui-plugin-examples
 
 > Note: Rancher Extensions is in early development - support for Extensions was recently added to Rancher 2.7.0 and the internal Rancher team is building out extensions using the extensions framework. Over time, the extensions API will improve and the supporting documentation will be built out, to enable the wider developer community to build their own extensions.
+
+
+// TODO: RC Add new section to cover IClusterProvisioner. Section should include cloud-crednetial and machine-config bits. Reference section from `code-base-works/machine-drivers` page
+// TODO: RC Add new info to addTab part about customParams
