+++ shell/core/types-provisioning.ts
@@ -1,4 +1,4 @@
-export type SaveHook = (hook: Function, name: string) => void;
+export type ClusterSaveHook = (hook: (cluster: any) => void, name: string, priority?: number, ) => void;
 
 /**
  * Interface that a custom Cluster Provisioner should implement
@@ -30,10 +30,10 @@ export interface IClusterProvisioner {
    * The `attributes: { kind: <value> }` should match the last part of the id
    * The `attributes: { group: <value> }` should match the remaining parts of the id
    */
-  machineConfigSchema: { [key: string]: any };
+  machineConfigSchema?: { [key: string]: any };
 
   /**
-   * Override the default method to create a machine config that will be inserted into a new machine pool
+   * Override the default method to create a machine config object that will be inserted into a new machine pool
    *
    * The machine config will be an instance related to the machine config schema
    *
@@ -41,9 +41,9 @@ export interface IClusterProvisioner {
    *
    * @param idx Index of new pool
    * @param pools Existing machine pools
-   * @returns TODO: RC
+   * @returns Instance of a machine config
    */
-  createMachinePoolMachineConfig(idx: number, pools: any[]): Promise<{[key: string]: any}>;
+  createMachinePoolMachineConfig?(idx: number, pools: any[]): Promise<{[key: string]: any}>;
 
   /**
    * Update the cluster before and or after the cluster is saved
@@ -56,7 +56,7 @@ export interface IClusterProvisioner {
    *  This allows the model received in response to the API request to be updated
    * @param cluster The cluster (`provisioning.cattle.io.cluster`)
    */
-  registerSaveHooks(registerBeforeHook: SaveHook, registerAfterHook: SaveHook, cluster: any): void;
+  registerSaveHooks?(registerBeforeHook: ClusterSaveHook, registerAfterHook: ClusterSaveHook, cluster: any): void;
 
   /**
    * Override the default process to save the machine config's associated with the machine pools
@@ -65,11 +65,14 @@ export interface IClusterProvisioner {
    *
    * The pool will have `create: true` if the pool is new or `update: true` if the pool already exists
    *
+   * For information on proxying HTTP requests from the browser via Rancher https://rancher.github.io/dashboard/code-base-works/machine-drivers#api-calls
+   * These docs also cover how to reference a Cloud Credential and use it's properties in the proxy's request `Authorization` header
+   *
    * @param pools Machine pools
    * @param cluster The cluster (`provisioning.cattle.io.cluster`)
-   * @returns TODO: RC
+   * @returns Content of async result / promise N/A, only the success / fail state
    */
-  saveMachinePoolConfigs(pools: any[], cluster: any): Promise<any>
+  saveMachinePoolConfigs?(pools: any[], cluster: any): Promise<any>
 
   /**
    * Existing tabs to show or hide in the cluster's detail view
@@ -110,9 +113,12 @@ export interface IClusterProvisioner {
   /**
    * Override the default process to save the cluster (`provisioning.cattle.io.cluster`)
    *
+   * For information on proxying HTTP requests from the browser via Rancher https://rancher.github.io/dashboard/code-base-works/machine-drivers#api-calls
+   * These docs also cover how to reference a Cloud Credential and use it's properties in the proxy's request `Authorization` header
+   *
    * @param pools Machine pools
    * @param cluster The cluster (`provisioning.cattle.io.cluster`)
    * @returns Array of errors. If there are no errors the array will be empty
    */
-  provision(cluster: any, pools: any[]): Promise<any[]>;
+  provision?(cluster: any, pools: any[]): Promise<any[]>;
 }
