+++ shell/store/plugins.js
@@ -59,9 +59,10 @@ const driverMap = {
 
 // Map a driver component back to the cloud credential field name their data has to be stored in
 const driverToFieldMap = {
-  aws:    'amazonec2',
-  gcp:    'google',
-  oracle: 'oci',
+  aws:                 'amazonec2',
+  gcp:                 'google',
+  oracle:              'oci',
+  digitaloceanexample: 'digitalocean' // TODO: RC DEBUG remove
 };
 
 // Machine driver fields that are probably a credential field
